requirements.txt
Flask-Migrate==2.6.0
Flask-Script
coverage
